chore: remove unused code and files
