+++ package.json
@@ -50,13 +50,11 @@
     "got": "^12.5.2",
     "if-env": "^1.0.4",
     "ky": "^0.32.1",
-    "little-state-machine": "^4.8.0",
     "react": "^18.2.0",
     "react-color": "^2.19.3",
     "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
     "react-hook-form": "^7.38.0",
-    "react-router-dom": "^6.4.2",
     "reactfire": "^4.2.2",
     "use-immer": "^0.7.0",
     "xstate": "^4.34.0",
